Extract session layout helper class
